Makes attractions and tabletop notifications use new task scheduler
