+++ requirements.txt
@@ -7,7 +7,7 @@ alembic==0.9.1
 treepoem==1.0.1
 email_validator==1.0.2
 phonenumbers==8.8.1
-pockets>=0.6.1
+pockets>=0.6.2
 residue>=0.2.4
 schedule>=0.5.0
 XlsxWriter==1.0.2
