+++ tests/uber/email_tests/email_fixtures.py
@@ -8,7 +8,7 @@ from uber.amazon_ses import AmazonSES
 from uber.automated_emails import AutomatedEmail
 from uber.config import c, Config
 from uber.models import Attendee
-from uber.tasks.emails import SendAllAutomatedEmailsJob
+from uber.tasks.email import SendAllAutomatedEmailsJob
 from uber.utils import DateBase, localize_datetime
